+++ main/signers/hot/SeedSigner/worker.ts
@@ -7,7 +7,7 @@ import {
   signMessage,
   signTransaction,
   signTypedData
-} from '../HotSigner/controller'
+} from '../HotSigner/worker'
 
 import type {
   CoreWorkerMethod,
